Cleanning data for Historical dataset and format the Fixture dataset
The historical and missing datasets, both obtained through web scraping (the historical dataset with BeautifulSoup and the missing dataset with Selenium), were merged to analyze the performance of both methods.

The Fixture dataset contains randomly generated values ​​and speculative data for the 30 countries already confirmed and the 18 countries that could potentially qualify for the final roster of teams for the 2026 FIFA World Cup. This random data was generated by me to visualize possible outcomes and practice my data science skills.